+++ TagCloud/RectanglesLayouter/PointsGenerator/SpiralPointsGenerator.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Drawing;
+using TagCloud.Data;
 
 namespace TagCloud.RectanglesLayouter.PointsGenerator
 {
